[Draft] Split Bezcurve from Draft.py and drop code from Draft Edit
[Draft] remove bezcurve code from draft
