+++ src/Mod/Draft/Draft.py
@@ -494,43 +494,9 @@ def makeBSpline(pointslist,closed=False,placement=None,face=None,support=None):
 
     return obj
 
-def makeBezCurve(pointslist,closed=False,placement=None,face=None,support=None,degree=None):
-    """makeBezCurve(pointslist,[closed],[placement]): Creates a Bezier Curve object
-    from the given list of vectors.   Instead of a pointslist, you can also pass a Part Wire."""
-    if not FreeCAD.ActiveDocument:
-        FreeCAD.Console.PrintError("No active document. Aborting\n")
-        return
-    if not isinstance(pointslist,list):
-        nlist = []
-        for v in pointslist.Vertexes:
-            nlist.append(v.Point)
-        pointslist = nlist
-    if placement: typecheck([(placement,FreeCAD.Placement)], "makeBezCurve")
-    if len(pointslist) == 2: fname = "Line"
-    else: fname = "BezCurve"
-    obj = FreeCAD.ActiveDocument.addObject("Part::Part2DObjectPython",fname)
-    _BezCurve(obj)
-    obj.Points = pointslist
-    if degree:
-        obj.Degree = degree
-    else:
-        import Part
-        obj.Degree = min((len(pointslist)-(1 * (not closed))),
-                         Part.BezierCurve().MaxDegree)
-    obj.Closed = closed
-    obj.Support = support
-    if face != None:
-        obj.MakeFace = face
-    obj.Proxy.resetcontinuity(obj)
-    if placement: obj.Placement = placement
-    if gui:
-        _ViewProviderWire(obj.ViewObject)
-#        if not face: obj.ViewObject.DisplayMode = "Wireframe"
-#        obj.ViewObject.DisplayMode = "Wireframe"
-        formatObject(obj)
-        select(obj)
-
-    return obj
+from draftobjects import bezcurve 
+makeBezCurve = bezcurve.make_bez_curve
+_BezCurve = bezcurve._BezCurve
 
 def makeText(stringslist,point=Vector(0,0,0),screen=False):
     """makeText(strings,[point],[screen]): Creates a Text object at the given point,
@@ -4785,133 +4751,6 @@ class _BSpline(_DraftObject):
 
 # for compatibility with older versions
 _ViewProviderBSpline = _ViewProviderWire
-
-class _BezCurve(_DraftObject):
-    """The BezCurve object"""
-
-    def __init__(self, obj):
-        _DraftObject.__init__(self,obj,"BezCurve")
-        obj.addProperty("App::PropertyVectorList","Points","Draft",QT_TRANSLATE_NOOP("App::Property","The points of the Bezier curve"))
-        obj.addProperty("App::PropertyInteger","Degree","Draft",QT_TRANSLATE_NOOP("App::Property","The degree of the Bezier function"))
-        obj.addProperty("App::PropertyIntegerList","Continuity","Draft",QT_TRANSLATE_NOOP("App::Property","Continuity"))
-        obj.addProperty("App::PropertyBool","Closed","Draft",QT_TRANSLATE_NOOP("App::Property","If the Bezier curve should be closed or not"))
-        obj.addProperty("App::PropertyBool","MakeFace","Draft",QT_TRANSLATE_NOOP("App::Property","Create a face if this curve is closed"))
-        obj.addProperty("App::PropertyLength","Length","Draft",QT_TRANSLATE_NOOP("App::Property","The length of this object"))
-        obj.addProperty("App::PropertyArea","Area","Draft",QT_TRANSLATE_NOOP("App::Property","The area of this object"))
-        obj.MakeFace = getParam("fillmode",True)
-        obj.Closed = False
-        obj.Degree = 3
-        obj.Continuity = []
-        #obj.setEditorMode("Degree",2)#hide
-        obj.setEditorMode("Continuity",1)#ro
-
-    def execute(self, fp):
-        self.createGeometry(fp)
-        fp.positionBySupport()
-
-    def _segpoleslst(self,fp):
-        """split the points into segments"""
-        if not fp.Closed and len(fp.Points) >= 2: #allow lower degree segment
-            poles=fp.Points[1:]
-        elif fp.Closed and len(fp.Points) >= fp.Degree: #drawable
-            #poles=fp.Points[1:(fp.Degree*(len(fp.Points)//fp.Degree))]+fp.Points[0:1]
-            poles=fp.Points[1:]+fp.Points[0:1]
-        else:
-            poles=[]
-        return [poles[x:x+fp.Degree] for x in \
-            range(0, len(poles), (fp.Degree or 1))]
-
-    def resetcontinuity(self,fp):
-        fp.Continuity = [0]*(len(self._segpoleslst(fp))-1+1*fp.Closed)
-        #nump= len(fp.Points)-1+fp.Closed*1
-        #numsegments = (nump // fp.Degree) + 1 * (nump % fp.Degree > 0) -1
-        #fp.Continuity = [0]*numsegments
-
-    def onChanged(self, fp, prop):
-        if prop == 'Closed': # if remove the last entry when curve gets opened
-            oldlen = len(fp.Continuity)
-            newlen = (len(self._segpoleslst(fp))-1+1*fp.Closed)
-            if oldlen > newlen:
-                fp.Continuity = fp.Continuity[:newlen]
-            if oldlen < newlen:
-                fp.Continuity = fp.Continuity + [0]*(newlen-oldlen)
-        if hasattr(fp,'Closed') and fp.Closed and prop in  ['Points','Degree','Closed'] and\
-                len(fp.Points) % fp.Degree: # the curve editing tools can't handle extra points
-            fp.Points=fp.Points[:(fp.Degree*(len(fp.Points)//fp.Degree))] #for closed curves
-        if prop in ["Degree"] and fp.Degree >= 1: #reset Continuity
-            self.resetcontinuity(fp)
-        if prop in ["Points","Degree","Continuity","Closed"]:
-            self.createGeometry(fp)
-
-    def createGeometry(self,fp):
-        import Part
-        plm = fp.Placement
-        if fp.Points:
-            startpoint=fp.Points[0]
-            edges = []
-            for segpoles in self._segpoleslst(fp):
-#                if len(segpoles) == fp.Degree # would skip additional poles
-                 c = Part.BezierCurve() #last segment may have lower degree
-                 c.increase(len(segpoles))
-                 c.setPoles([startpoint]+segpoles)
-                 edges.append(Part.Edge(c))
-                 startpoint = segpoles[-1]
-            w = Part.Wire(edges)
-            if fp.Closed and w.isClosed():
-                try:
-                    if hasattr(fp,"MakeFace"):
-                        if fp.MakeFace:
-                            w = Part.Face(w)
-                    else:
-                        w = Part.Face(w)
-                except Part.OCCError:
-                    pass
-            fp.Shape = w
-            if hasattr(fp,"Area") and hasattr(w,"Area"):
-                fp.Area = w.Area
-            if hasattr(fp,"Length") and hasattr(w,"Length"):
-                fp.Length = w.Length            
-        fp.Placement = plm
-
-    @classmethod
-    def symmetricpoles(cls,knot, p1, p2):
-        """make two poles symmetric respective to the knot"""
-        p1h=FreeCAD.Vector(p1)
-        p2h=FreeCAD.Vector(p2)
-        p1h.multiply(0.5)
-        p2h.multiply(0.5)
-        return ( knot+p1h-p2h , knot+p2h-p1h)
-
-    @classmethod
-    def tangentpoles(cls,knot, p1, p2,allowsameside=False):
-        """make two poles have the same tangent at knot"""
-        p12n=p2.sub(p1)
-        p12n.normalize()
-        p1k=knot-p1
-        p2k=knot-p2
-        p1k_= FreeCAD.Vector(p12n)
-        kon12=(p1k*p12n)
-        if allowsameside or not (kon12 < 0 or p2k*p12n > 0):# instead of moving
-            p1k_.multiply(kon12)
-            pk_k=knot-p1-p1k_
-            return (p1+pk_k,p2+pk_k)
-        else:
-            return cls.symmetricpoles(knot, p1, p2)
-
-    @staticmethod
-    def modifysymmetricpole(knot,p1):
-        """calculate the coordinates of the opposite pole
-        of a symmetric knot"""
-        return knot+knot-p1
-
-    @staticmethod
-    def modifytangentpole(knot,p1,oldp2):
-        """calculate the coordinates of the opposite pole
-        of a tangent knot"""
-        pn=knot-p1
-        pn.normalize()
-        pn.multiply((knot-oldp2).Length)
-        return pn+knot
 
 # for compatibility with older versions ???????
 _ViewProviderBezCurve = _ViewProviderWire
