+++ src/Mod/Draft/draftobjects/baseobject.py
@@ -1,0 +1,49 @@
+"""This module provides the base code for Draft Objects.
+At the moment the class is a duplicate of Draft._DraftObject
+"""
+## @package baseobject
+# \ingroup DRAFT
+# \brief This module provides the code for Draft Base Object.
+
+# ***************************************************************************
+# *                                                                         *
+# *   This file is part of the FreeCAD CAx development system.              *
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU Lesser General Public License (LGPL)    *
+# *   as published by the Free Software Foundation; either version 2 of     *
+# *   the License, or (at your option) any later version.                   *
+# *   for detail see the LICENCE text file.                                 *
+# *                                                                         *
+# *   FreeCAD is distributed in the hope that it will be useful,            *
+# *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+# *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+# *   GNU Library General Public License for more details.                  *
+# *                                                                         *
+# *   You should have received a copy of the GNU Library General Public     *
+# *   License along with FreeCAD; if not, write to the Free Software        *
+# *   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+# *   USA                                                                   *
+# *                                                                         *
+# ***************************************************************************
+
+
+class _DraftObject:
+    """The base class for Draft objects"""
+    def __init__(self,obj,tp="Unknown"):
+        if obj:
+            obj.Proxy = self
+        self.Type = tp
+
+    def __getstate__(self):
+        return self.Type
+
+    def __setstate__(self,state):
+        if state:
+            self.Type = state
+
+    def execute(self,obj):
+        pass
+
+    def onChanged(self, obj, prop):
+        pass
